Added scraper to operate on a list of pages
Pages are provided to the script as a file where each line is a page
name. The output is a file containing a valid JSON array with each
post being an object. In order to avoid hangups, if a page access
fails the scraping for that page is aborted. Some code has also been
changed to prevent crashes from bad inputs or to just generally
clean up.